rename folder net to model
